backend/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
torch
torchvision
transformers
Pillow
